+++ src/app/docs/inputs/demo-radio/demo-radio.component.ts
@@ -63,7 +63,7 @@ export class DemoRadioComponent implements OnInit {
         RadioModule,
     ],
 })
-export class DemoToggleModule {
+export class DemoRadioModule {
 }`,
                     },
                 },