+++ src/app/docs/layout/demo-divider/demo-divider.component.ts
@@ -1,4 +1,5 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
+import { DemoPageModel } from '../../../shared/model';
 
 @Component({
     selector: 'app-demo-divider',
@@ -7,10 +8,41 @@ import { Component, OnInit } from '@angular/core';
 })
 export class DemoDividerComponent implements OnInit {
 
+    @ViewChild('defaultRef', {static: true}) defaultRef: any;
+
+    page: DemoPageModel;
+
     constructor() {
     }
 
     ngOnInit(): void {
+        this.page = {
+            title: 'Divider component example',
+            subtitle: 'It may seem useless, but still usable',
+            demos: [
+                {
+                    title: 'Default usage',
+                    description: '',
+                    templateRef: this.defaultRef,
+                    values: {
+                        html: '<sdk-divider></sdk-divider>',
+                        module: `import { DividerModule } from 'ngx-slice-kit';
+
+@NgModule({
+    imports: [
+        CommonModule,
+        FormsModule,
+        // add DividerModule export
+        DividerModule,
+    ],
+})
+export class DemoDividerModule {
+}`
+                    },
+                }
+            ],
+            apis: [],
+        };
     }
 
 }