+++ src/app/docs/layout/layout.module.ts
@@ -17,6 +17,7 @@ import { DemoPaginationComponent } from './demo-pagination/demo-pagination.compo
 import { DemoProgressComponent } from './demo-progress/demo-progress.component';
 import { DemoSlidesComponent } from './demo-slides/demo-slides.component';
 import { DemoTableComponent } from './demo-table/demo-table.component';
+import { SharedModule } from "../../shared/shared.module";
 
 const entryComponents = [];
 
@@ -41,6 +42,7 @@ const moduleDeclarations = [
         CommonModule,
         DemoLayoutRoutingModule,
         SliceKitModule,
+        SharedModule,
     ],
     declarations: moduleDeclarations,
     exports: moduleDeclarations,